fix: memory leak from route cache holding dead client handles
- Clear disconnected client from route cache entries
  Stale cache entries held Arc<ClientWriteHandle> refs, preventing
  WriteBuffer cleanup even after client disconnect and keepalive expiry

- Enforce keep-alive config (default_keep_alive, max_keep_alive)
  Previously clients with keep_alive=0 would never timeout

- Detect dead sockets via EPOLLHUP/EPOLLERR/close events
  Immediately disconnect when epoll reports socket errors

- Fix clippy warnings in packet.rs and client.rs

+++ crates/mqlite-core/src/packet.rs
@@ -1543,7 +1543,7 @@ pub fn encode_subscribe(subscribe: &Subscribe, buf: &mut Vec<u8>) {
         let opts_byte = (options.qos as u8)
             | if options.no_local { 0x04 } else { 0 }
             | if options.retain_as_published { 0x08 } else { 0 }
-            | ((options.retain_handling as u8) << 4);
+            | (options.retain_handling << 4);
         payload.push(opts_byte);
     }
